backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
neo4j==5.28.1
python-dotenv==1.0.1
faker==37.1.0
